requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# src/modules/buildModel.py: 8
Keras == 2.3.1

# src/modules/buildModel.py: 1
feather_format == 0.4.0

# src/modules/preProcess.py: 6,7,8
# src/modules/word2vec.py: 1
nltk == 3.3

# src/modules/preProcess.py: 2
numpy == 1.18.1

# src/modules/preProcess.py: 1
pandas == 0.23.4

# src/modules/preProcess.py: 9,10
spacy == 2.2.3

# src/modules/buildModel.py: 2,9,10,11
tensorflow == 1.14.0
